Add Media Query and Scrolling Divs
Added a media query to make phone image move down when adjusting the
browser width. Added a similar query for a tablet breakpoint (768px).
Added a mobile query which re-arranges and adjusts the content

Added buttons that enable a user to scroll through information
to find out more about what the app can do.

TO-DO:
* Figure out positioning of scrolling divs so when the content
change it doesn't shift the content of the page
* Figure out why the `Sign Up` button intermittently loses its hover
  effect ONLY when you hover over the actual text
* Add animations for changing divs and buttons on mobile